Add test class for taxi
